# batches: Faster workspace lookup