--- a/migrations/frontend/1655763641/down.sql
+++ b/migrations/frontend/1655763641/down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS batch_spec_workspaces_id_batch_spec_id;