--- a/migrations/frontend/1655763641/metadata.yaml
+++ b/migrations/frontend/1655763641/metadata.yaml
@@ -1,0 +1,3 @@
+name: faster_workspace_batch_spec_lookup
+parents: [1655481894]
+createIndexConcurrently: true