--- a/migrations/frontend/1655763641/up.sql
+++ b/migrations/frontend/1655763641/up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY IF NOT EXISTS batch_spec_workspaces_id_batch_spec_id ON batch_spec_workspaces(id, batch_spec_id);