--- a/migrations/frontend/squashed.sql
+++ b/migrations/frontend/squashed.sql
@@ -3480,6 +3480,8 @@ CREATE INDEX batch_spec_workspace_execution_jobs_user_id ON batch_spec_workspace
 
 CREATE INDEX batch_spec_workspaces_batch_spec_id ON batch_spec_workspaces USING btree (batch_spec_id);
 
+CREATE INDEX batch_spec_workspaces_id_batch_spec_id ON batch_spec_workspaces USING btree (id, batch_spec_id);
+
 CREATE INDEX batch_specs_rand_id ON batch_specs USING btree (rand_id);
 
 CREATE INDEX changeset_jobs_bulk_group_idx ON changeset_jobs USING btree (bulk_group);