@eseliger 2022-06-21
If you compare the two query plans from the PR description, adding the new index makes it so that postgres can find records matching on id and batch spec id in one index lookup, without looping over ~600000 records and fetching the batch spec id value from the heap. Also, the index stores data in an ordered fashion so getting the first 50 is cheaper by that as well because it can just look at the first few in the index until it collected 50 records. 

I am bad at explaining this, hope it makes any sense :D